+++ src/main/java/com/example/rest_controllers/StillingController.java
@@ -1,11 +1,13 @@
 package com.example.rest_controllers;
 
 import com.example.database_classes.Stilling;
+import com.example.security.TokenManager;
 import com.example.sql_folder.SqlQueries;
 import com.fasterxml.jackson.databind.util.JSONPObject;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.*;
 
+import javax.security.auth.message.AuthException;
 import javax.websocket.server.PathParam;
 
 /**
@@ -16,24 +18,43 @@ import javax.websocket.server.PathParam;
 public class StillingController {
     SqlQueries query = new SqlQueries();
     @RequestMapping("/stilling/{beskrivelse}")
-    public Stilling getStilling(@PathVariable("beskrivelse") String beskrivelse) {
-        Stilling ret = query.selectStilling(beskrivelse);
-        return query.selectStilling(beskrivelse);
+    public Stilling getStilling(@PathVariable("beskrivelse") String beskrivelse, @RequestHeader (value = "token") String token) throws AuthException {
+        if (TokenManager.verifiser(token)) {
+            return query.selectStilling(beskrivelse);
+        }
+        else {
+            throw new AuthException("Token not authenticated");
+        }
     }
 
     @RequestMapping(value="/stilling/delete", method=RequestMethod.POST)
-    public boolean deleteStilling(@RequestBody Stilling stilling) {
-        return query.deleteStilling(stilling);
+    public boolean deleteStilling(@RequestBody Stilling stilling, @RequestHeader (value = "token") String token) throws AuthException {
+        if (TokenManager.verifiser(token)) {
+            return query.deleteStilling(stilling);
+        }
+        else {
+            throw new AuthException("Token not authenticated");
+        }
     }
 
     @RequestMapping(value="/stilling/add", method=RequestMethod.POST)
-    public boolean insertStilling(@RequestBody Stilling stilling) {
-        return query.insertStilling(stilling);
+    public boolean insertStilling(@RequestBody Stilling stilling, @RequestHeader (value = "token") String token) throws AuthException {
+        if (TokenManager.verifiser(token)) {
+            return query.insertStilling(stilling);
+        }
+        else {
+            throw new AuthException("Token not authenticated");
+        }
     }
 
     @RequestMapping(value="/stilling/update", method=RequestMethod.POST)
-    public boolean updateStilling(@RequestBody Stilling stilling){
-        return query.updateStilling(stilling);
+    public boolean updateStilling(@RequestBody Stilling stilling, @RequestHeader (value = "token") String token) throws AuthException {
+        if (TokenManager.verifiser(token)) {
+            return query.updateStilling(stilling);
+        }
+        else {
+            throw new AuthException("Token not authenticated");
+        }
     }
 }
